+++ src/main/java/com/project/TFIBackendSpringBoot/service/PatientServiceImpl.java
@@ -4,6 +4,8 @@ package com.project.TFIBackendSpringBoot.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.TFIBackendSpringBoot.dto.PatientDTO;
 import com.project.TFIBackendSpringBoot.dto.PatientDTOSave;
+import com.project.TFIBackendSpringBoot.exceptions.ResourseAlreadyExistsExeption;
+import com.project.TFIBackendSpringBoot.exceptions.ResourseNotFoundException;
 import com.project.TFIBackendSpringBoot.model.Patient;
 import com.project.TFIBackendSpringBoot.repository.IPatientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,44 +36,75 @@ public class PatientServiceImpl implements IPatientService<PatientDTO, PatientDT
     }
 
     @Override
-    public void save(PatientDTOSave patientDTOSave) {
-        savePatient(patientDTOSave);
+    public void save(PatientDTOSave patientDTOSave)throws ResourseAlreadyExistsExeption {
+        Patient patient=patientRepository.findByDNI(patientDTOSave.getDNI());
+
+        if (patient==null){
+
+            savePatient(patientDTOSave);
+
+        }else{
+
+            throw new ResourseAlreadyExistsExeption("Patient with DNI: " + patientDTOSave.getDNI() + " is already loaded in database");
+
+        }
+
     }
 
     @Override
-    public void remove(String dni) {
-        try{
+    public void remove(String dni)throws ResourseNotFoundException {
 
             Patient patient=patientRepository.findByDNI(dni);
-            patientRepository.deleteById(patient.getId());
 
-        }catch (RuntimeException e){
-            System.out.println("Errorico");
-            System.out.println(e);
-        }
+            if (patient==null){
+                throw new ResourseNotFoundException("Patient with DNI: " + dni + " doesn't exist in database");
+            }else{
+
+                patientRepository.deleteById(patient.getId());
+
+            }
+
+
+
     }
 
     @Override
-    public PatientDTO search(String dni) {
-
+    public PatientDTO search(String dni)throws ResourseNotFoundException {
 
         Patient patient= patientRepository.findByDNI(dni);
 
-        PatientDTO patientDTO=mapper.convertValue(patient, PatientDTO.class);
+        if (patient==null){
+            throw new ResourseNotFoundException("Patient with DNI: " + dni + " doesn't exist in database");
+        }else{
+
+            PatientDTO patientDTO=mapper.convertValue(patient, PatientDTO.class);
+
+            return patientDTO;
+
+        }
 
 
-        return patientDTO;
+
     }
 
     @Override
-    public Set<PatientDTO> searchAll() {
+    public Set<PatientDTO> searchAll()throws ResourseNotFoundException {
+
         List<Patient> patients= patientRepository.findAll();
+
+        if (patients.isEmpty()){
+            throw new ResourseNotFoundException("Patients list is empty");
+        }else{
+
         Set<PatientDTO> patientsDTO=new HashSet<>();
         for (Patient patient:patients) {
             PatientDTO patientDTO=mapper.convertValue(patient,PatientDTO.class);
             patientsDTO.add(patientDTO);
         }
-        return patientsDTO;
+
+            return patientsDTO;
+
+        }
     }
 
     @Override
